Implement url encode and decode functions for objects
